--- a/attribute/benchmark_test.go
+++ b/attribute/benchmark_test.go
@@ -272,7 +272,7 @@ func BenchmarkStringSlice(b *testing.B) {
 	b.Run("Emit", benchmarkEmit(kv))
 }
 
-func BenchmarkDistinctMapLookup(b *testing.B) {
+func BenchmarkDistinctMap(b *testing.B) {
 	const size = 10000
 	m := make(map[attribute.Distinct]struct{}, size)
 	keys := make([]attribute.Distinct, size)
@@ -282,27 +282,19 @@ func BenchmarkDistinctMapLookup(b *testing.B) {
 		keys[i] = s.Equivalent()
 	}
 
-	var out struct{}
+	b.Run("Set", func(b *testing.B) {
+		for n := 0; n < b.N; n++ {
+			m[keys[n%size]] = struct{}{}
+		}
+	})
 
-	b.ResetTimer()
-	for n := 0; n < b.N; n++ {
-		out = m[keys[n%size]]
-	}
+	b.Run("Lookup", func(b *testing.B) {
+		var out struct{}
 
-	_ = out
-}
+		for n := 0; n < b.N; n++ {
+			out = m[keys[n%size]]
+		}
 
-func BenchmarkDistinctMapSet(b *testing.B) {
-	const size = 10000
-	m := make(map[attribute.Distinct]struct{}, size)
-	keys := make([]attribute.Distinct, size)
-	for i := 0; i < size; i++ {
-		s := attribute.NewSet(attribute.Int("key", i))
-		keys[i] = s.Equivalent()
-	}
-
-	b.ResetTimer()
-	for n := 0; n < b.N; n++ {
-		m[keys[n%size]] = struct{}{}
-	}
+		_ = out
+	})
 }