--- a/attribute/hash.go
+++ b/attribute/hash.go
@@ -10,6 +10,25 @@ import (
 	"go.opentelemetry.io/otel/attribute/internal/fnv"
 )
 
+// Type identifiers. These identifiers are hashed before the value of the
+// corresponding type. This is done to distinguish values that are hashed with
+// the same value representation (e.g. `int64(1)` and `true`, []int64{0} and
+// int64(0)).
+//
+// These are all 8 byte length strings converted to a uint64 representation. A
+// uint64 is used instead of the string directly as an optimization, it avoids
+// the for loop in [fnv] which adds minor overhead.
+const (
+	boolID         uint64 = 7953749933313450591 // "_boolean" (little endian)
+	int64ID        uint64 = 7592915492740740150 // "64_bit_i" (little endian)
+	float64ID      uint64 = 7376742710626956342 // "64_bit_f" (little endian)
+	stringID       uint64 = 6874584755375207263 // "_string_" (little endian)
+	boolSliceID    uint64 = 6875993255270243167 // "_[]bool_" (little endian)
+	int64SliceID   uint64 = 3762322556277578591 // "_[]int64" (little endian)
+	float64SliceID uint64 = 7308324551835016539 // "[]double" (little endian)
+	stringSliceID  uint64 = 7453010373645655387 // "[]string" (little endian)
+)
+
 // hashKVs returns a new FNV-1a hash of kvs.
 func hashKVs(kvs []KeyValue) fnv.Hash {
 	h := fnv.New()
@@ -25,42 +44,38 @@ func hashKV(h fnv.Hash, kv KeyValue) fnv.Hash {
 
 	switch kv.Value.Type() {
 	case BOOL:
-		h = h.String("b")
+		h = h.Uint64(boolID)
 		h = h.Uint64(kv.Value.numeric)
 	case INT64:
-		h = h.String("i")
+		h = h.Uint64(int64ID)
 		h = h.Uint64(kv.Value.numeric)
 	case FLOAT64:
-		h = h.String("f")
+		h = h.Uint64(float64ID)
 		// Assumes numeric stored with math.Float64bits.
 		h = h.Uint64(kv.Value.numeric)
 	case STRING:
-		h = h.String("s")
+		h = h.Uint64(stringID)
 		h = h.String(kv.Value.stringly)
 	case BOOLSLICE:
-		// Differentiate between bool and [1]bool.
-		h = h.String("[]b")
+		h = h.Uint64(boolSliceID)
 		rv := reflect.ValueOf(kv.Value.slice)
 		for i := 0; i < rv.Len(); i++ {
 			h = h.Bool(rv.Index(i).Bool())
 		}
 	case INT64SLICE:
-		// Differentiate between int64 and [1]int64.
-		h = h.String("[]i")
+		h = h.Uint64(int64SliceID)
 		rv := reflect.ValueOf(kv.Value.slice)
 		for i := 0; i < rv.Len(); i++ {
 			h = h.Int64(rv.Index(i).Int())
 		}
 	case FLOAT64SLICE:
-		// Differentiate between float64 and [1]float64.
-		h = h.String("[]f")
+		h = h.Uint64(float64SliceID)
 		rv := reflect.ValueOf(kv.Value.slice)
 		for i := 0; i < rv.Len(); i++ {
 			h = h.Float64(rv.Index(i).Float())
 		}
 	case STRINGSLICE:
-		// Differentiate between string and [1]string.
-		h = h.String("[]s")
+		h = h.Uint64(stringSliceID)
 		rv := reflect.ValueOf(kv.Value.slice)
 		for i := 0; i < rv.Len(); i++ {
 			h = h.String(rv.Index(i).String())