This presents a proof-of-concept for how to replace the internals of `Distinct` with a hash of the data held by a `Set`.

### Why do this?

Currently, the `Set` and `Distinct` both hold the same data: an interface with that `KeyValue` array.

The Go runtime has [optimized code paths](https://github.com/golang/go/blob/9e3b1d53a012e98cfd02de2de8b1bd53522464d4/src/runtime/alg.go#L19) for many built-in types when they are used as a map key. Our current design, using a comparable `interface{}` pointing to an array, [is not as optimized](https://github.com/golang/go/blob/9e3b1d53a012e98cfd02de2de8b1bd53522464d4/src/runtime/alg.go#L168-L179) as one of these built-in types.

By replacing the `Distinct` type with a single hash field (i.e. `~uint64`), we can realize performance imporvements without giving up any of the current `Set` or `Distinct` behaviors.

```console
$ benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/attribute
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
                     │   old.txt   │               new.txt               │
                     │   sec/op    │   sec/op     vs base                │
DistinctMap/Set-8      96.30n ± 4%   22.41n ± 1%  -76.74% (p=0.000 n=10)
DistinctMap/Lookup-8   92.16n ± 3%   20.20n ± 1%  -78.09% (p=0.000 n=10)
geomean                94.21n        21.27n       -77.42%
```

This is all done by computing the `Set`s data hash when it is created. This adds minimal overhead to the set creation.

```console
benchstat old_NewSet.txt new_NewSet.txt
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/attribute
cpu: Intel(R) Core(TM) i7-8550U CPU @ 1.80GHz
         │ old_NewSet.txt │           new_NewSet.txt            │
         │     sec/op     │   sec/op     vs base                │
NewSet-8      233.2n ± 4%   266.7n ± 1%  +14.34% (p=0.000 n=10)

         │ old_NewSet.txt │         new_NewSet.txt         │
         │      B/op      │    B/op     vs base            │
NewSet-8       64.00 ± 0%   64.00 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

         │ old_NewSet.txt │         new_NewSet.txt         │
         │   allocs/op    │ allocs/op   vs base            │
NewSet-8       1.000 ± 0%   1.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```

Given the performance costs (`NewSet`) are generally paid once for many look-ups using that `Set` (i.e. the metric SDK), and the cost is recouped in the outsized improvement in just one setting/lookup of a map, this change seems like a positive net performance improvement.

### Why not do this?

The hashing algorithm used to compute the `Set` data hash is the [fnv-1a hash](https://en.wikipedia.org/wiki/Fowler%E2%80%93Noll%E2%80%93Vo_hash_function#FNV-1a_hash).  This hash was chosen for a few reasons:

- Supports 64 bit values
- Fast
- Relatively low collision rate
- Easy [reference implementation](https://pkg.go.dev/hash/fnv) in Go

The [`hash/fnv`](https://pkg.go.dev/hash/fnv) implementation is copied locally and updated to not require any allocations when it is used. This means that it can be checked using the same tests (and even verified by an independent implementation if needed).

This all sounds good, but there is one big issue. This has a low collision rate, but the current implementation does not have any collision rate. Meaning by accepting these changes we will add a very small probability of invalid data setting/look-ups.