@pellared 2024-03-06
> This all sounds good, but there is one big issue. This has a low collision rate, but the current implementation does not have any collision rate. Meaning by accepting these changes we will add a very small probability of invalid data setting/look-ups.

I think that this trade-off is acceptable because:
- low performance overhead is a very important property for instrumentation, if I understand correctly, this change will increase the performance of the metrics SDK
- for telemetry, the harm caused by a collision should be acceptable