@MadVikingGod 2024-03-08
Where I would expect to see the real gains are in the metric aggregation benchmarks.
```
cd sdk/metric/internal/aggregate
go test -bench='.*/.*/.*/100'
```

I do worry that most users will use `WithAttributes` vs `WithAttributeSet`, which could erase these gains.  But, I don't think we have a benchmark that measures the difference.